warning fix
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@2186 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Files/preprocess_map_shared.cpp
@@ -110,7 +110,7 @@ make_nonconflicting_filename_variant(/*const*/ FileSpecifier& inBaseName, FileSp
                 else
                         theBaseNameLengthToUse = theBaseNameLength;
                 
-                sprintf(theNameToTryString, "%.*s%s", theBaseNameLengthToUse, theBaseNameString, theVariantSuffix);
+                sprintf(theNameToTryString, "%.*s%s", (int) theBaseNameLengthToUse, theBaseNameString, theVariantSuffix);
         
                 theNameToTry.FromDirectory(theDirectory);
 #if defined(mac) || defined(SDL_RFORK_HACK)
